third_party/blink/web_tests/fast/shapes/parsing/parsing-shape-lengths-expected.txt

Test that basic shapes accept different length units

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getStyleValue("shape-outside", "circle(0 at 0 0)") is "circle(0px at 0% 0%)"
PASS getStyleValue("shape-outside", "circle(1px at +1px -1px)") is "circle(1px at 1px -1px)"
PASS getStyleValue("shape-outside", "circle(1.5px at -1.5px +1.5px)") is "circle(1.5px at -1.5px 1.5px)"
PASS getStyleValue("shape-outside", "circle(.5px at -.5px +.5px)") is "circle(0.5px at -0.5px 0.5px)"
PASS getStyleValue("shape-outside", "inset(1cm 1mm 1in 1px round 1pt 1pc)") is "inset(1cm 1mm 1in 1px round 1pt 1pc)"
PASS getStyleValue("shape-outside", "inset(-1px 1px 1px 1px round 1px 1px 1px 1px)") is "inset(-1px 1px 1px round 1px)"
PASS getStyleValue("shape-outside", "inset(1px -1px 1px 1px round 1px 1px 1px 1px)") is "inset(1px -1px 1px 1px round 1px)"
PASS getStyleValue("shape-outside", "inset(1px 1px -1px 1px round 1px 1px 1px 1px)") is "inset(1px 1px -1px round 1px)"
PASS getStyleValue("shape-outside", "inset(1px 1px 1px -1px round 1px 1px 1px 1px)") is "inset(1px 1px 1px -1px round 1px)"
PASS getStyleValue("shape-outside", "circle(1em at -1em +1em)") is "circle(1em at -1em 1em)"
PASS getStyleValue("shape-outside", "circle(1.5em at -1.5em +1.5em)") is "circle(1.5em at -1.5em 1.5em)"
PASS getStyleValue("shape-outside", "circle(.5em at -.5em +.5em)") is "circle(0.5em at -0.5em 0.5em)"
PASS getStyleValue("shape-outside", "circle(1ex at 1ex 1ex)") is "circle(1ex at 1ex 1ex)"
PASS getStyleValue("shape-outside", "circle(1rem at 1rem 1rem)") is "circle(1rem at 1rem 1rem)"
PASS getComputedStyleValue("shape-outside", "circle(1.5em at .5em 1em)") is "circle(12px at 4px 8px)"
PASS getComputedStyleValue("shape-outside", "circle(1.5rem at .5rem 1rem)") is "circle(24px at 8px 16px)"
PASS getStyleValue("shape-outside", "circle(1vw at 1vw 1vw)") is "circle(1vw at 1vw 1vw)"
PASS getStyleValue("shape-outside", "circle(1vh at 1vh 1vh)") is "circle(1vh at 1vh 1vh)"
PASS getStyleValue("shape-outside", "circle(1vmin at 1vmin 1vmin)") is "circle(1vmin at 1vmin 1vmin)"
PASS getComputedStyleValue("shape-outside", "circle(1.5vw at .5vw 1vw)") is "circle(12px at 4px 8px)"
PASS getComputedStyleValue("shape-outside", "circle(1.5vh at .5vh 1vh)") is "circle(9px at 3px 6px)"
PASS getComputedStyleValue("shape-outside", "circle(1.5vmin at .5vmin 1vmin)") is "circle(9px at 3px 6px)"
PASS getStyleValue("shape-outside", "circle(100% at 100% 100%)") is "circle(100% at 100% 100%)"
PASS getStyleValue("shape-outside", "inset(45% 45% 90% 60% round 25% 10%)") is "inset(45% 45% 90% 60% round 25% 10%)"
PASS getStyleValue("shape-outside", "ellipse(100% 100% at 100% 100%)") is "ellipse(100% 100% at 100% 100%)"
PASS getStyleValue("shape-outside", "polygon(10% 20%, 30% 40%, 40% 50%)") is "polygon(10% 20%, 30% 40%, 40% 50%)"
PASS getComputedStyleValue("shape-outside", "circle(150% at 50% 100%)") is "circle(150% at 50% 100%)"
PASS getComputedStyleValue("shape-outside", "inset(45% 45% 90% 60% round 25% 10%)") is "inset(45% 45% 90% 60% round 25% 10%)"
PASS getComputedStyleValue("shape-outside", "ellipse(100% 100% at 100% 100%)") is "ellipse(100% 100% at 100% 100%)"
PASS getComputedStyleValue("shape-outside", "polygon(10% 20%, 30% 40%, 40% 50%)") is "polygon(10% 20%, 30% 40%, 40% 50%)"
PASS getStyleValue("shape-outside", "circle(1 at 1px 1px)") is ""
PASS getComputedStyleValue("shape-outside", "circle(1 at 1px 1px)") is "none"
PASS getStyleValue("shape-outside", "circle(px at 1px 1px)") is ""
PASS getComputedStyleValue("shape-outside", "circle(px at 1px 1px)") is "none"
PASS getStyleValue("shape-outside", "circle(1p at 1px 1px)") is ""
PASS getComputedStyleValue("shape-outside", "circle(1p at 1px 1px)") is "none"
PASS getStyleValue("shape-outside", "circle(calc( at 1px 1px))") is ""
PASS getComputedStyleValue("shape-outside", "circle(calc( at 1px 1px))") is "none"
PASS getStyleValue("shape-outside", "circle(-1.5px at -1.5px +1.5px)") is ""
PASS getComputedStyleValue("shape-outside", "circle(-1.5px at -1.5px +1.5px)") is "none"
PASS getStyleValue("shape-outside", "inset(1cm 1mm 1in 1px round 1pt -1pc)") is ""
PASS getComputedStyleValue("shape-outside", "inset(1cm 1mm 1in 1px round 1pt -1pc)") is "none"
PASS getStyleValue("shape-outside", "ellipse(-1em 1em at 1em 1em)") is ""
PASS getComputedStyleValue("shape-outside", "ellipse(-1em 1em at 1em 1em)") is "none"
PASS getStyleValue("shape-outside", "ellipse(1em -1em at 1em 1em)") is ""
PASS getComputedStyleValue("shape-outside", "ellipse(1em -1em at 1em 1em)") is "none"
PASS successfullyParsed is true

TEST COMPLETE
